UPSTREAM: ASoC: SOF: Intel: hda: modify stream interrupt handler
Modify the stream interrupt handler to always wake up the
IRQ thread if the status register is valid. The IRQ thread
performs the check for stream interrupts and RIRB interrupts
in a loop to handle the case of missed interrupts when an
unsolicited response from the codec is received just before the
stream interrupt handler is completed.

Signed-off-by: Kai Vehmanen <kai.vehmanen@linux.intel.com>
Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>
Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 6297a0dc4c14a62bea5a9137ceef280cb7a80665)

BUG=b:987654321
TEST=Test Audio use cases for CML with full SOF patch series applied.

Change-Id: Ib78f3d67ce749cbc343bfe9852225979d0e367ce
Signed-off-by: Ap, Kamal <kamal.ap@intel.corp-partner.google.com>